LibCore+LibGUI: Move DesktopServices to LibCore
