+++ Libraries/LibCore/Makefile
@@ -2,6 +2,7 @@ OBJS = \
     ArgsParser.o \
     ConfigFile.o \
     DateTime.o \
+    DesktopServices.o \
     DirIterator.o \
     ElapsedTimer.o \
     Event.o \
