+++ Libraries/LibGUI/Makefile
@@ -18,7 +18,6 @@ OBJS = \
     CppLexer.o \
     CppSyntaxHighlighter.o \
     Desktop.o \
-    DesktopServices.o \
     Dialog.o \
     DisplayLink.o \
     DragOperation.o \
